Refactor repository tests.
